Split multi-strategy provider files into per-strategy files
Split providers with multiple strategies so each strategy lives in its
own file. The main file keeps the provider struct, Meta, Auth, init, and
shared helpers.

claude/:
  claude.go  - provider struct, Meta, Auth, saveClaudeCredential, init
  oauth.go   - OAuthStrategy, credential loading, token refresh, CLI refresh, parse
  apikey.go  - APIKeyStrategy, loadAPIKey
  web.go     - WebStrategy, session loading, org lookup, parse

kimicode/:
  kimicode.go - provider struct, Meta, Auth, init, shared helpers, fetchUsage, parse
  device.go   - DeviceFlowStrategy, device code polling, RunDeviceFlow
  apikey.go   - APIKeyStrategy, loadAPIKey

gemini/:
  gemini.go  - provider struct, Meta, Auth, init, shared helpers
  oauth.go   - OAuthStrategy, credential loading, quota fetch, parse
  apikey.go  - APIKeyStrategy, loadAPIKey, model validation

amp/:
  amp.go         - provider struct, Meta, Auth, init, loadCLISecretsToken,
                   fetchBalance, parseDisplayBalance, RPC types
  clisecrets.go  - CLISecretsStrategy
  apikey.go      - APIKeyStrategy

Same code, same line count, just better file-level intent. No new
abstractions, no import changes, no behavioral differences.
